Mar 19, 2024, 5:20 AM

+++ helpers.h
@@ -18,9 +18,9 @@ struct Region {
 template<typename T>
 void insertionSort(std::vector<T>& vec)
 {
-  for (int i = 1; i < vec.size(); ++i) {
+  for (size_t i = 1; i < vec.size(); ++i) {
     T key = vec[i];
-    int j = i;
+    size_t j = i;
     while (j > 0 && key < vec[j - 1]) {
       vec[j] = vec[j - 1];
       --j;
